Rename `ListItemIDP` into `ModuleIDP`
